Remove problematic "negative-value" and redundant tests
ICU seems to have changed the default formatting for negative numbers in the
en-US locale, and since that really isn't testing any feature of Money, we can
remove those tests without loss of coverage. We can also remove the de-DE test
which was not improving coverage above what was already there.